requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
pydantic-settings
